+++ front/src/applications/editor/tools/rangeEdition/speedSection/LpvPanelCard.tsx
