+++ front/src/applications/editor/tools/rangeEdition/speedSection/LpvPanelSubSection.tsx
