+++ front/src/applications/editor/tools/toolTypes.ts
@@ -2,6 +2,7 @@ enum TOOL_TYPES {
   SELECTION = 'select-items',
   TRACK_EDITION = 'track-edition',
   SPEED_SECTION_EDITION = 'speed-edition',
+  CATENARY_EDITION = 'catenary',
   SWITCH_EDITION = 'switch-edition',
   SIGNAL_EDITION = 'signal-edition',
   DETECTOR_EDITION = 'detector-edition',
