+++ front/src/applications/editor/tools/tools.ts
@@ -4,7 +4,7 @@ import {
   DetectorEditionTool,
   SignalEditionTool,
 } from './pointEdition/tools';
-import { SpeedEditionTool } from './rangeEdition/tools';
+import { CatenaryEditionTool, SpeedEditionTool } from './rangeEdition/tools';
 import TOOL_TYPES from './toolTypes';
 import SelectionTool from './selection/tool';
 import SwitchEditionTool from './switchEdition/tool';
@@ -14,6 +14,7 @@ const TOOLS = Object.freeze({
   [TOOL_TYPES.SELECTION]: SelectionTool,
   [TOOL_TYPES.TRACK_EDITION]: TrackEditionTool,
   [TOOL_TYPES.SPEED_SECTION_EDITION]: SpeedEditionTool,
+  [TOOL_TYPES.CATENARY_EDITION]: CatenaryEditionTool,
   [TOOL_TYPES.SWITCH_EDITION]: SwitchEditionTool,
   [TOOL_TYPES.SIGNAL_EDITION]: SignalEditionTool,
   [TOOL_TYPES.DETECTOR_EDITION]: DetectorEditionTool,
