+++ front/src/types/editor.ts
@@ -1,12 +1,7 @@
 import { JSONSchema7 } from 'json-schema';
 import { Feature, GeoJsonProperties, Geometry, Point, LineString, MultiLineString } from 'geojson';
 
-import {
-  Direction,
-  ObjectType,
-  DirectionalTrackRange,
-  TrackRange,
-} from '../common/api/osrdEditoastApi';
+import { Direction, ObjectType, DirectionalTrackRange } from '../common/api/osrdEditoastApi';
 import { NullGeometry } from './geospatial';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
